+++ libgui/FileProgress.cpp
@@ -37,7 +37,7 @@
 //***************************************************************************
 FileProgress::FileProgress(QWidget *parent,
 	const QUrl &url, unsigned int size,
-	unsigned int samples, unsigned int rate, unsigned int bits,
+	unsigned int samples, double rate, unsigned int bits,
 	unsigned int tracks)
     :QSemiModal(parent, "FileProgress", true),
     m_url(url),
